Clean up partitioned output functionality
