+++ packages/op-datasets/src/op_datasets/etl/ingestion/sinks.py
@@ -4,11 +4,11 @@ import polars as pl
 from op_coreutils import duckdb_local as utilsduckdb
 from op_coreutils.clickhouse import insert_arrow
 from op_coreutils.logger import structlog
+from op_coreutils.partitioned import SinkMarkerPath, WrittenParquetPath
 from op_coreutils.storage.gcs_parquet import (
     gcs_upload_parquet,
     local_upload_parquet,
 )
-from op_coreutils.storage.paths import PartitionedOutput, SinkMarkerPath
 
 from op_datasets.etl.ingestion.markers import IngestionCompletionMarker
 
@@ -21,14 +21,14 @@ log = structlog.get_logger()
 class RawOnchainDataSink:
     location: RawOnchainDataLocation
 
-    def write_single_part(self, dataframe: pl.DataFrame, part_output: PartitionedOutput):
+    def write_single_part(self, dataframe: pl.DataFrame, part_output: WrittenParquetPath):
         if self.location == RawOnchainDataLocation.GCS:
-            gcs_upload_parquet(part_output.path.full_path, dataframe)
+            gcs_upload_parquet(part_output.full_path, dataframe)
             return
 
         elif self.location == RawOnchainDataLocation.LOCAL:
             local_upload_parquet(
-                path=self.location.with_prefix(part_output.path.full_path),
+                path=self.location.with_prefix(part_output.full_path),
                 df=dataframe,
             )
